Scrapped the dynamic column names in return tibble.
I realised that having column names in the returned tibble
with changing column names would make it impossible for
users to experiment with different r-values without having
to also change downstream R code (e.g., in ggplots).

The only reasonable approach is thus to have the tauc() function
return predefined column names. To avoid ambiguities I chose
to simple use x and y (x is the energy scale, as input, y is
the calculated absorbance to the power of 1/r).